move contextual analyzer to SonarTS and make it a Sensor. Server, now renamed to ContextualServer, is not starting anymore
